[#2] refactor: split stub with unit test
